requirements.txt
discord.py
requests
python-decouple
asyncio